Separate into components and move to new folder
