+++ packages/block-editor/package.json
@@ -45,6 +45,7 @@
 		"classnames": "^2.2.5",
 		"diff": "^3.5.0",
 		"dom-scroll-into-view": "^1.2.1",
+		"fast-average-color": "4.3.0",
 		"inherits": "^2.0.3",
 		"lodash": "^4.17.15",
 		"memize": "^1.0.5",
