+++ packages/block-editor/src/components/background-controls/index.js
@@ -1,0 +1,1 @@
+export { default as __experimentalWithBackgroundControls } from './with-background-controls';
