+++ packages/block-editor/src/components/background-controls/shared.js
@@ -1,8 +1,7 @@
 export const IMAGE_BACKGROUND_TYPE = 'image';
 export const VIDEO_BACKGROUND_TYPE = 'video';
-export const ALLOWED_MEDIA_TYPES = [ IMAGE_BACKGROUND_TYPE, VIDEO_BACKGROUND_TYPE ];
-
 export const COVER_MIN_HEIGHT = 50;
+export const ALLOWED_MEDIA_TYPES = [ IMAGE_BACKGROUND_TYPE, VIDEO_BACKGROUND_TYPE ];
 
 export function backgroundImageStyles( url ) {
 	return url ?
